api/pyproject.toml
[project]
name = "start-up-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "argon2-cffi>=23.1.0",
    "chapa>=0.1.2",
    "fastapi[standard]>=0.115.12",
    "git-filter-repo>=2.47.0",
    "phonenumbers>=9.0.2",
    "psycopg2>=2.9.10",
    "pydantic-extra-types>=2.10.3",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
]
